- make enemy select rune - make rune overlays behave as expected - update rune models with correct better shading

+++ src/World/Models/Runes/BabiRune.js
@@ -6,16 +6,18 @@ export default class BabiRune extends Rune {
   beforeResolution() {}
 
   afterResolution() {
-    // const attackerPlayer = this.experience.world.getPlayerAtTurn()
-    const defenderPlayer = this.experience.world.getPlayerAtTurn(false)
-    const tier = this.rune[this.selectedTier]
-    this.payTierPrice()
-    if (this.didPayTierPrice) {
-      setTimeout(() => {
-        /* maybe make vfx here to show life stones are beeing destroyed */
-        defenderPlayer.destroyLifeStones(tier.value)
-      }, 800)
-      this.didPayTierPrice = false
+    if (this.owner.selectedRune) {
+      // const attackerPlayer = this.experience.world.getPlayerAtTurn()
+      const defenderPlayer = this.experience.world.getPlayerAtTurn(false)
+      const tier = this.rune[this.owner.selectedRune?.tier]
+      this.payTierPrice()
+      if (this.didPayTierPrice) {
+        setTimeout(() => {
+          /* maybe make vfx here to show life stones are beeing destroyed */
+          defenderPlayer.destroyLifeStones(tier.value)
+        }, 800)
+        this.didPayTierPrice = false
+      }
     }
   }
 }
